feat(elreal): widen block exponent to integer<256> per McCleeary's unbounded-exponent design
McCleeary's LFPERA carries the per-block exponent in Z (Haskell Integer,
unbounded). The streaming division (dissertation 4.2.6) recurses with
newdiv = g0*divisor each level, so the divisor's magnitude -- and its block
exponent -- roughly DOUBLES per level. A fixed int32 (or even int64) exponent
overflows after ~11 (~59) levels, capping multi-block division precision. We
were reformulating Ryan's proven algorithm to fit int32; that is the wrong
approach.

Follow the dissertation instead: carry the exponent in the library's fixed-size
integer<256, uint32_t>. It is a trivially-copyable POD, so block<FpType> remains
trivially copyable / destructible -- the #925 hardware-shareable invariant the
construction test enforces -- while its ~2^255 range never overflows for any
realizable host (it outlasts even a quad host's denormal floor by orders of
magnitude). This realizes the spirit of Ryan's unbounded exponent without
giving up the trivial block layout.

Changes:
- block.hpp: add block<FpType>::exp_t = integer<256,uint32_t>; exp and
  exponent() become exp_t; createZero takes exp_t; value_as narrows for ldexp.
- threeAdd.hpp / block_eft.hpp: alignment and product/quotient exponents flow
  as exp_t (auto); ldexp shift args narrowed to int (host-range alignment).
- sum.hpp / divide.hpp / math/exponent.hpp: eager + math paths keep int (their
  blocks are host-range) with explicit narrowing of exponent().
- elreal_oracle.hpp + block/eft/threeAdd tests: narrow exp_t to int for the
  dyadic oracle / ldexp (host-range verification values).

Validated: full elreal suite 34/34 PASS on gcc and clang; block construction
triviality static_asserts still hold.

Relates to #1061

+++ elastic/elreal/arithmetic/exact_value_oracle.cpp
@@ -88,12 +88,12 @@ using namespace sw::universal::elreal_oracle;
 template <typename FpType>
 double dval(const block<FpType>& b) {
     if (b.is_zero_block()) return 0.0;
-    return std::ldexp(static_cast<double>(b.v), b.exp);
+    return std::ldexp(static_cast<double>(b.v), static_cast<int>(b.exp));
 }
 
 template <typename FpType>
 double block_ulp(const block<FpType>& b) {
-    const int E = b.is_zero_block() ? b.exp : b.exponent();
+    const int E = static_cast<int>(b.is_zero_block() ? b.exp : b.exponent());
     return std::ldexp(1.0, E - (block<FpType>::k - 1));
 }
